% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cumulant.R
\name{fourth_order_cumulant}
\alias{fourth_order_cumulant}
\title{Calculate Fourth order Autocumulant}
\usage{
fourth_order_cumulant(x, lag)
}
\arguments{
\item{x}{Sample}

\item{lag}{The lags at which to calculate cumulants, a vector of length 3}
}
\description{
Calculate Fourth order Autocumulant
}
