% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cumulant.R
\name{general_cumulant}
\alias{general_cumulant}
\title{Autocumulant Calculation (second, third or fourth)}
\usage{
general_cumulant(x, lag)
}
\arguments{
\item{x}{Sample}

\item{lag}{Lag vector}

\item{iter}{Number of Monte Carlo Simulations}
}
\description{
Autocumulant Calculation (second, third or fourth)
}
