% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_predict.R
\name{quad_predict}
\alias{quad_predict}
\title{Calculate Linear or Quadratic Predictor}
\usage{
quad_predict(x, k.pred = 1, past = 20, type = "lin")
}
\arguments{
\item{x}{Vector from GARCH process}

\item{k.pred}{Number of future points to be predicted}

\item{type}{Type of Prediction, has to be quad or lin}

\item{iter}{Number of iterations}
}
\description{
Calculate Linear or Quadratic Predictor
}
