% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cumulant.R
\name{second_order_cumulant}
\alias{second_order_cumulant}
\title{Calculate Second order Autocumulant}
\usage{
second_order_cumulant(x, lag)
}
\arguments{
\item{x}{Sample from which autocumulant to be calculated}

\item{lag}{The lags at which to calculate cumulants, a scalar}
}
\description{
Calculate Second order Autocumulant
}
