% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cumulant.R
\name{third_order_cumulant}
\alias{third_order_cumulant}
\title{Calculate Third order Autocumulant}
\usage{
third_order_cumulant(x, lag)
}
\arguments{
\item{x}{Sample from which autocumulant to be calculated}

\item{lag}{The lags at which to calculate cumulants, a vector of length 2}
}
\description{
Calculate Third order Autocumulant
}
